crna_solution_challenge/src/app/backend/CRNA_Submission_Run/requirements.txt
Flask
gunicorn
google-cloud-bigquery
google-cloud-pubsub